pyproject.toml
[tool.poetry]
name = "py-clash-configer"
version = "0.1.0"
description = "Merge Clash configuration subscription and local overrides."
authors = ["genzj <zj0512@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.26.0"
PyYAML = "^6.0"
click = "^8.0.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
